+++ libs/pages/elewa/invest/src/lib/pages-elewa-invest.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { InvestPageComponent } from './pages/invest-page/invest-page.component';
 
-import { InvestRoutingModule } from './pages/invest.routing';
+import { InvestRoutingModule } from './invest.routing';
 
 @NgModule({
   imports: [CommonModule, InvestRoutingModule],
